Add gorm

+++ mock/repository_mock.go
@@ -19,7 +19,7 @@ func (r *RepositoryMock) CreateUser(user model.User) (string, error) {
 	return r.Email, r.Err
 }
 
-func (r *RepositoryMock) GetUser(id float64) (model.User, error) {
+func (r *RepositoryMock) GetUser(id uint) (model.User, error) {
 	return r.User, r.Err
 }
